move util sql classes
