+++ src/main/java/com/mercari/solution/module/transform/BeamSQLTransform.java
@@ -6,9 +6,9 @@ import com.mercari.solution.util.cloud.google.ParameterManagerUtil;
 import com.mercari.solution.util.cloud.google.StorageUtil;
 import com.mercari.solution.util.schema.RowSchemaUtil;
 import com.mercari.solution.util.schema.converter.ElementToRowConverter;
-import com.mercari.solution.util.sql.udf.AggregateFunctions;
-import com.mercari.solution.util.sql.udf.ArrayFunctions;
-import com.mercari.solution.util.sql.udf.MathFunctions;
+import com.mercari.solution.util.domain.sql.beamsql.udf.AggregateFunctions;
+import com.mercari.solution.util.domain.sql.beamsql.udf.ArrayFunctions;
+import com.mercari.solution.util.domain.sql.beamsql.udf.MathFunctions;
 import org.apache.beam.sdk.coders.RowCoder;
 import org.apache.beam.sdk.extensions.sql.SqlTransform;
 import org.apache.beam.sdk.transforms.DoFn;
