+++ src/main/java/com/mercari/solution/util/domain/sql/beamsql/udf/MathFunctions.java
